Refactor encryption and compression logic, update paths
